Split Job Engine Application modules to improve code reusage
