+++ job-engine/app/resources/src/main/resources/openapi/jobEngine-scopeId-jobId.yaml
