--- job-engine/app/src/main/java/org/eclipse/kapua/job/engine/app/converter/KapuaIdParamConverterProvider.java
@@ -1,52 +1,0 @@
-/*******************************************************************************
- * Copyright (c) 2021 Eurotech and/or its affiliates and others
- *
- * This program and the accompanying materials are made
- * available under the terms of the Eclipse Public License 2.0
- * which is available at https://www.eclipse.org/legal/epl-2.0/
- *
- * SPDX-License-Identifier: EPL-2.0
- *
- * Contributors:
- *     Eurotech - initial API and implementation
- *******************************************************************************/
-package org.eclipse.kapua.job.engine.app.converter;
-
-import java.lang.annotation.Annotation;
-import java.lang.reflect.Type;
-
-import javax.ws.rs.ext.ParamConverter;
-import javax.ws.rs.ext.ParamConverterProvider;
-import javax.ws.rs.ext.Provider;
-
-import org.eclipse.kapua.locator.KapuaLocator;
-import org.eclipse.kapua.model.id.KapuaId;
-import org.eclipse.kapua.model.id.KapuaIdFactory;
-
-@Provider
-public class KapuaIdParamConverterProvider implements ParamConverterProvider {
-
-    private final KapuaLocator locator = KapuaLocator.getInstance();
-    private final KapuaIdFactory kapuaIdFactory = locator.getFactory(KapuaIdFactory.class);
-
-    @Override
-    public <T> ParamConverter<T> getConverter(Class<T> rawType, Type genericType, Annotation[] annotations) {
-        if (rawType == KapuaId.class) {
-            return new ParamConverter<T>() {
-
-                @Override
-                public T fromString(String value) {
-                    return rawType.cast(kapuaIdFactory.newKapuaId(value));
-                }
-
-                @Override
-                public String toString(T value) {
-                    return ((KapuaId) value).toCompactId();
-                }
-
-            };
-        }
-        return null;
-    }
-
-}
