--- job-engine/app/src/main/java/org/eclipse/kapua/job/engine/app/jackson/deserializer/KapuaIdDeserializer.java
@@ -1,36 +1,0 @@
-/*******************************************************************************
- * Copyright (c) 2021 Eurotech and/or its affiliates and others
- *
- * This program and the accompanying materials are made
- * available under the terms of the Eclipse Public License 2.0
- * which is available at https://www.eclipse.org/legal/epl-2.0/
- *
- * SPDX-License-Identifier: EPL-2.0
- *
- * Contributors:
- *     Eurotech - initial API and implementation
- *******************************************************************************/
-package org.eclipse.kapua.job.engine.app.jackson.deserializer;
-
-import java.io.IOException;
-
-import org.eclipse.kapua.locator.KapuaLocator;
-import org.eclipse.kapua.model.id.KapuaId;
-import org.eclipse.kapua.model.id.KapuaIdFactory;
-
-import com.fasterxml.jackson.core.JsonParser;
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.DeserializationContext;
-import com.fasterxml.jackson.databind.JsonDeserializer;
-
-public class KapuaIdDeserializer extends JsonDeserializer<KapuaId> {
-
-    private final KapuaLocator locator = KapuaLocator.getInstance();
-    private final KapuaIdFactory kapuaIdFactory = locator.getFactory(KapuaIdFactory.class);
-
-    @Override
-    public KapuaId deserialize(JsonParser p, DeserializationContext ctxt) throws IOException, JsonProcessingException {
-        return kapuaIdFactory.newKapuaId(p.getValueAsString());
-    }
-
-}
