--- job-engine/app/src/main/java/org/eclipse/kapua/job/engine/app/jackson/mixin/IsJobRunningResponseMixin.java
@@ -1,22 +1,0 @@
-/*******************************************************************************
- * Copyright (c) 2021 Eurotech and/or its affiliates and others
- *
- * This program and the accompanying materials are made
- * available under the terms of the Eclipse Public License 2.0
- * which is available at https://www.eclipse.org/legal/epl-2.0/
- *
- * SPDX-License-Identifier: EPL-2.0
- *
- * Contributors:
- *     Eurotech - initial API and implementation
- *******************************************************************************/
-package org.eclipse.kapua.job.engine.app.jackson.mixin;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-
-public interface IsJobRunningResponseMixin {
-
-    @JsonProperty("isRunning")
-    boolean isRunning();
-
-}
