--- job-engine/app/src/main/java/org/eclipse/kapua/job/engine/app/jackson/mixin/JobStartOptionsMixin.java
@@ -1,22 +1,0 @@
-/*******************************************************************************
- * Copyright (c) 2021 Eurotech and/or its affiliates and others
- *
- * This program and the accompanying materials are made
- * available under the terms of the Eclipse Public License 2.0
- * which is available at https://www.eclipse.org/legal/epl-2.0/
- *
- * SPDX-License-Identifier: EPL-2.0
- *
- * Contributors:
- *     Eurotech - initial API and implementation
- *******************************************************************************/
-package org.eclipse.kapua.job.engine.app.jackson.mixin;
-
-import org.eclipse.kapua.job.engine.jbatch.JobStartOptionsImpl;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
-
-@JsonDeserialize(as = JobStartOptionsImpl.class)
-@JsonIgnoreProperties("type")
-public interface JobStartOptionsMixin { }
