--- job-engine/app/src/main/resources/shiro.ini
@@ -1,53 +1,0 @@
-# =======================
-# Shiro INI configuration
-# =======================
-
-[main]
-# Objects and their properties are defined here,
-# Such as the securityManager, Realms and anything
-# else needed to build the SecurityManager
-
-#authenticator
-authenticator = org.eclipse.kapua.service.authentication.shiro.KapuaAuthenticator
-securityManager.authenticator = $authenticator
-
-#
-# Auth filters
-rebuildSessionFilter = org.eclipse.kapua.job.engine.app.filter.RebuildSessionFilter
-
-##########
-# Realms #
-##########
-# Login
-kapuaUserPassAuthenticatingRealm = org.eclipse.kapua.service.authentication.shiro.realm.UserPassAuthenticatingRealm
-kapuaApiKeyAuthenticatingRealm = org.eclipse.kapua.service.authentication.shiro.realm.ApiKeyAuthenticatingRealm
-kapuaJwtAuthenticatingRealm = org.eclipse.kapua.service.authentication.shiro.realm.JwtAuthenticatingRealm
-
-# Session
-kapuaAccessTokenAuthenticatingRealm = org.eclipse.kapua.service.authentication.shiro.realm.AccessTokenAuthenticatingRealm
-
-# Authorization
-kapuaAuthorizingRealm = org.eclipse.kapua.service.authorization.shiro.KapuaAuthorizingRealm
-
-securityManager.realms = $kapuaAuthorizingRealm, $kapuaAccessTokenAuthenticatingRealm, $kapuaApiKeyAuthenticatingRealm, $kapuaUserPassAuthenticatingRealm, $kapuaJwtAuthenticatingRealm
-
-# 90*24*60*60 seconds = 90 days = 7776000 seconds
-securityManager.rememberMeManager.cookie.name = kapua-rememberme
-securityManager.rememberMeManager.cookie.maxAge = 0
-
-[users]
-# The 'users' section is for simple deployments
-# when you only need a small number of statically-defined
-# set of User accounts.
-
-[roles]
-# The 'roles' section is for simple deployments
-# when you only need a small number of statically-defined
-# roles.
-
-[urls]
-# The 'urls' section is used for url-based security
-# in web applications.  We'll discuss this section in the
-# Web documentation
-
-/v1/** = rebuildSessionFilter
