--- job-engine/app/src/main/webapp/WEB-INF/web.xml
@@ -1,87 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-    Copyright (c) 2021 Eurotech and/or its affiliates and others
-
-    This program and the accompanying materials are made
-    available under the terms of the Eclipse Public License 2.0
-    which is available at https://www.eclipse.org/legal/epl-2.0/
-
-    SPDX-License-Identifier: EPL-2.0
-
-    Contributors:
-        Eurotech - initial API and implementation
- -->
-
-<!-- This web.xml file is not required when using Servlet 3.0 container,
-    see implementation details http://jersey.java.net/nonav/documentation/latest/jax-rs.html -->
-<web-app xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xmlns="http://java.sun.com/xml/ns/javaee"
-         xsi:schemaLocation="http://java.sun.com/xml/ns/javaee http://java.sun.com/xml/ns/javaee/web-app_3_0.xsd"
-         id="WebApp_ID" version="3.0">
-
-    <display-name>Kapua Job Engine REST API</display-name>
-
-    <!-- -->
-    <!-- Listeners -->
-    <listener>
-        <listener-class>org.apache.shiro.web.env.EnvironmentLoaderListener</listener-class>
-    </listener>
-
-    <listener>
-        <listener-class>org.eclipse.kapua.job.engine.app.JobEngineRestApiListener</listener-class>
-    </listener>
-
-    <!-- -->
-    <!-- Filters -->
-
-    <filter>
-        <filter-name>CORSResponseFilter</filter-name>
-        <filter-class>org.eclipse.kapua.app.api.core.CORSResponseFilter</filter-class>
-    </filter>
-    <filter-mapping>
-        <filter-name>CORSResponseFilter</filter-name>
-        <url-pattern>/v1/*</url-pattern>
-    </filter-mapping>
-
-    <filter>
-        <filter-name>ShiroFilter</filter-name>
-        <filter-class>org.apache.shiro.web.servlet.ShiroFilter</filter-class>
-    </filter>
-    <filter-mapping>
-        <filter-name>ShiroFilter</filter-name>
-        <url-pattern>/*</url-pattern>
-    </filter-mapping>
-
-    <filter>
-        <filter-name>KapuaSessionCleanUpFilter</filter-name>
-        <filter-class>org.eclipse.kapua.app.api.core.filter.KapuaSessionCleanupFilter</filter-class>
-    </filter>
-    <filter-mapping>
-        <filter-name>KapuaSessionCleanUpFilter</filter-name>
-        <url-pattern>/v1/*</url-pattern>
-    </filter-mapping>
-
-    <!-- -->
-    <!-- Servlets -->
-    <servlet>
-        <servlet-name>Jersey Web Application</servlet-name>
-        <servlet-class>org.glassfish.jersey.servlet.ServletContainer</servlet-class>
-        <init-param>
-            <param-name>javax.ws.rs.Application</param-name>
-            <param-value>org.eclipse.kapua.job.engine.app.JobEngineApplication</param-value>
-        </init-param>
-       <load-on-startup>1</load-on-startup>
-    </servlet>
-    <servlet-mapping>
-        <servlet-name>Jersey Web Application</servlet-name>
-        <url-pattern>/v1/*</url-pattern>
-    </servlet-mapping>
-
-    <session-config>
-        <session-timeout>1</session-timeout>
-        <cookie-config>
-            <max-age>0</max-age>
-        </cookie-config>
-    </session-config>
-
-</web-app>
